Do not use dispatch class in data_collection.cpp
data_collection.cpp stores information of kernel groups in a format
that assumes the number of dispatch classes.  This assumption is
unnecessary and inflexible.

Also, it seems RecordKernelGroups were never called, so the info is missing
in the dump.  I connected the logic in program.cpp.
Now it shows both kernel group data and dispatch data in the dump:

```
Program 1: Ran 1 time(s).
	HalProgrammableCoreType::TENSIX Kernel Groups: 2
		{TENSIX_DM:4 } on cores {[(x=0,y=0) - (x=0,y=0)]}
		{TENSIX_DM:4 TENSIX_DM:5 TENSIX_COMPUTE:6 } on cores {[(x=0,y=0) - (x=0,y=0)]}
	data_collector_t::DISPATCH_DATA_BINARY stats:
	  TENSIX_DM_0 binary data:
		max_transaction_size = 580
		min_transaction_size = 580
		num_writes           = 1
		total_write_size     = 580
		transaction_counts   = [580:1 ]
...
```