Didt test hang N300 (#27084)
### Ticket
Failing test:
[Link to Github
Issue](https://github.com/tenstorrent/tt-metal/actions/runs/17032526819)

### Problem description
N300 version of test runs on all chips (2), and is more sensitive to
Didt. Test is failing due to hang causing timeout.

### What's changed
The sdxl tests should be called with TT_MM_THROTTLE_PERF=5 since that
was the lowest level of throttle found to help by the customer team, and
that is the level they use to run the model.

### Checklist
- [x] Passing L2 Nightly
https://github.com/tenstorrent/tt-metal/actions/runs/17103477421/