Adding deallocate_input and reallocate_halo_output functionality to pools (#27007)
### Ticket
[27004](https://github.com/tenstorrent/tt-metal/issues/27004)
[19406](https://github.com/tenstorrent/tt-metal/issues/19406)

### Problem description
Pool ops consist of two micro ops: halo and micro pool op.
Halo is producing its own output which is needed from pool micro op but
is in some way a duplicate of input that is already taking up space. If
original input of pool op is not used by any succeeding op it is useful
to have a bool that indicates if the original input can be deallocated.
This bool will be called deallocate_input and will be set to false by
default. Usually deallocating original input tensor produces memory
fragmentation which can be resolved by moving halo output tesnor after
deallocation of the original input. Whether to do so or not will be
indicated by bool reallocate_halo_output which is used ideally with
deallocate_input set to true.

### What's changed
Added conditional deallocate of the input and conditional movement of
the halo output depending on the bools that are set. Due to the fact
that all the paths of the pool op are merging into one halo call, there
was a need for only two ifs in the right place in generic pools.

YOLOV8x: 
I haven't updated the input to the next max pool correctly in one of the
previous PRs this was hiding a bug where we are doing the maxpool 3
times on conv output instead chaining the output of previous maxpool as
input of the next one, this has now showed since we deallocate sharded
copies of inputs that are initially in DRAM by default, which showed
that inputs weren't propagated correctly.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/17152738006)
- [x] [Nightly tt-metal L2
tests](https://github.com/tenstorrent/tt-metal/actions/runs/17048054839)
- [x] [(Blackhole) Blackhole nightly
tests](https://github.com/tenstorrent/tt-metal/actions/runs/17048059635)
- [x] [(Single-card) Frequent model and ttnn tests
](https://github.com/tenstorrent/tt-metal/actions/runs/17128385805)
- [x] [(Single-card) Device perf
regressions](https://github.com/tenstorrent/tt-metal/actions/runs/17128399039)