#0: Fix distributed host buffer allocations (#27302)
### Ticket
N/A

### Problem description

https://github.com/tenstorrent/tt-metal/commit/196f240c45283b47a8f0b853da8fc5728ab1dbe6#diff-34de76e6a46483e94eb0ee1d3c41fb97539596ec5df778a48c8b2032dc6dc74c
introduced an optimization to avoid re-processing of replicated data
shards. It broke allocation code, as it relied on 2 step process to
create empty shards, and then allocate them to non-empty buffers in
parallel.

### What's changed
Fix the problematic code.

### Checklist
- [x] [T3K
tests](https://github.com/tenstorrent/tt-metal/actions/runs/17158054923)
- [X] New/Existing tests provide coverage for changes - tested locally
that the regression is gone