Change DeepSeek to route to minimal all gather (#27475)
### Ticket
#27377

### Problem description
DeepSeek routes to the command processor all gather which does not have
all the optimizations or handshaking that minimal all gather does,
resulting in a race and poor performance.

### What's changed
Use two semaphores instead of one, which will route you to the final all
gather. Create a different set of semaphores for the point to point
operation.

### TODO
Part of our plan going forward is to delete the command processor
version and internalize the semaphore creation. We don't want to
internalize the semaphore creation to a deprecated implementation, so we
want to delete the command processor version. To do that, we need to
first remove it from the codebase.

### Checklist
- [ ] DeepSeek Pipelines:
https://github.com/tenstorrent/tt-metal/actions/runs/17254517296