bugprone-reserved-identifier (#27519)
### Ticket
#22758 
Closes #27516 

### Problem description
We had this check disabled, because we are not respecting the rules set
forth in C++ standard.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/reserved-identifier.html

### What's changed
- Enable the check for the majority of the repo
- Fix some violations

Please note that I am leaving the check disabled for `llrt`, because
there are so many violations of the rules in `tt_metal/hw/inc`.
Correcting them would take a lot of effort, for little benefit.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/17276847937)
CI passes
- [x] New/Existing tests provide coverage for changes