Create max pool with indices LLK on SFPU (#606)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/17202

### Problem description
Maxpool op that also return the index of the max value is needed. Topk
local sort can accomplish this, but also performs a lot of extra work
(sorting whole tile column) while here we only need to extract single
max value.

### What's changed
Wrote implementation for max reduce with indices for 9 row tiles. This
is the specific case that is needed to meet perf requirement for a
customer.
Other cases/more general implementation can be added as needed, and
based on this one (just needs to be expanded).

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

### Checklist
- no tests in tt-llk yet, tested on tt-metal branch which will be merged
with a test