Add better default chunks_per_sync and num_workers_per_link for all_gather_minimal (#27603)
### Ticket
#27438

### Problem description
Currently we simply default these two hyperparameters to max chunks and
1 workers, which is generally worse. Ideally we have some derivable
formula for such a thing.

### What's changed
After sweeping through a list of each of these values, I found a formula
that always beats the current defaults. There are several values that do
so within standard deviation of each other, so I picked a formula that's
generally best, with the hyperparameters still there if we want to tune
them.

Findings:
- Chunks_per_sync of 1,2,5 caused a steep drop off. 10-20 were generally
slightly below 40-320, with 160 offering the best balance by a very
small margin. Used a ReLU160 chunks_per_sync calculation that is either
the total number of packets on that link up to 160.
- 1 worker was generally bad compared to 2-8 workers. 4 was generally
the best with 2 a close second. 8 only matched 4 and we would run out of
cores on a 4 link system anyway. Used a search strategy where we try to
maximize the number of cores we use given that we have enough in our
subdevice. If we have enough cores for 4 per link, use 4, otherwise try
2, otherwise try 1.


BW Comparisons, Old vs New on DiT shapes

|Topology | Old Defaults Average BW (GB/s) | New Defaults Average BW
(GB/s) |
| --- | --- | --- |
| Linear | 5.03 | 8.45 |
| Ring | 4.02 | 5.91 |


### TODO
- The last parameter num_channels_per_buffer made such a small
difference that I excluded it from this sweep for now, we will attempt
this later.
- Also need to do this for RS.


### Checklist
- [ ] T3K:
https://github.com/tenstorrent/tt-metal/actions/runs/17304449669
- [ ] 6U:
https://github.com/tenstorrent/tt-metal/actions/runs/17346872672