[skip ci] Allow user to choose to tag image as latest in release container wrapper (#27310)
Currently the publish-release-image-wrapper script allows devs to upload
latest-rc tagged Docker images (specifically release and
release-models), while only major new Metal releases push to latest.
This is in line with desired behavior, but in certain cases where the
containers need a quick update (such as when we introduced
release-models), it may be beneficial to bypass this process and push
the built images directly to latest. This PR introduces that option.

---------

Signed-off-by: June Knauth (tt) <jknauth@tenstorrent.com>
Co-authored-by: Raymond Kim <109366641+tt-rkim@users.noreply.github.com>