[skip ci] bugprone-nondeterministic-pointer-iteration-order (#27888)
### Ticket
#22758 
Closes #27887 

### Problem description
We had this check disabled.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/nondeterministic-pointer-iteration-order.html
Iterating over an unordered map of pointers occurs in nondeterministic
order.

If iteration order is irrelevant this can be ignored, but it should be
considered during development.

```
When violations are basically fine

You truly don’t care about order 
(e.g., you’re just visiting all items, and the order never leaks into user-visible output, hashing, or scheduling decisions). 
In such cases, the warning is informational. The docs say the check is advisory for exactly this reason.
```

### What's changed
- Enable check
- Disable existing violations (PLEASE CHECK THEM)