Fix SFP_STOCH_RND instruction modifiers (#639)
### Ticket
none

### Problem description
SFP_STOCH_RND modes were being used that relied on one of the opcode
bits being ignored in current HW

### What's changed
Bit 3 is ignored in these opcodes, so remap 14 -> 6 and 9 -> 1 to be
consistent with usage elsewhere in the SW, with the enums that are
defined for this modifier, and with the tt-isa-documentation specs.

### Type of change

- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

### Checklist
<!-- These are required steps and need to be run from tt-metal
repository's Actions. Use links below and replace them with your run -->
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)