build: create slim ird image for llk testing only (#640)
### Ticket
None

### Problem description
The LLK IRD image is based on Metal's IRD, so we could run Metal and LLK
testing with one image. The devil is in the details. The LLK IRD image
includes 2.5 GB of data that we don't need when running the LLK testing
framework (this comes from Metal). Instead of pulling 2.5 GB every time
we make a reservation, we can now pull 514 MB in a fraction of the time.

### What's changed
New, slim IRD image has been added.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update