fix some technically incorrect code (#28038)
While inspecting the code, I noticed we were inconsistently handling TXQ
ID variables when checking for eth TXQ busy. This isn't a problem in
main because all fabric variants only enable a single txq anyways, so
all the variables end up aliasing to the same thing regardless. This
will mostly be a problem when we flip the switch to enable 2-erisc mode
for fabric on Blackhole

### Checklist
- [x] All post commit - fast fabric tests only:
https://github.com/tenstorrent/tt-metal/actions/runs/17502015979