[POOL] Incorrect pool output when number of channels % 32 is in interval (16,32) (#27832)
### Ticket
[27738](https://github.com/tenstorrent/tt-metal/issues/27738)

### Problem description
The tilize_reconfig optimization is disabled when last_tile_is_partial
is true because partial tiles can contain junk data that needs to be
cleared.
For 280 channels which is the case mentioned in the issue, the partial
tile which is 24 channels is larger than FACE_WIDTH, so it spans 2 faces
and requires careful handling.

### What's changed
There was an issue where the bool flag for the partial tiles was
incorrectly calculated. So when channel size in last tile is > 16 and
<=32 this was considered as non partial tile. Where this should be
considered as partial tile just the number of output faces which last
tile produces should be 2 (same as if the tile was whole). This was
changed to reflect it correctly.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/17462034976)
- [x] [Nightly tt-metal L2
tests](https://github.com/tenstorrent/tt-metal/actions/runs/17433853977)
- [x] [(Single-card) Device perf
regressions](https://github.com/tenstorrent/tt-metal/actions/runs/17433836043)