Add default workers and chunks_per_sync values for reduce_scatter_minimal_async based on a performance sweep (#28196)
### Ticket
#27438

### Problem description
RS minimal did a lazy default to 1 worker and one sync. There was
potential for a more intelligent worker and sync scheme.

### What's changed

- Swept across sync values of [10, 20, 40, 80, 160, DEFAULT]
- Swept across workers per link of [8, 4, 2, 1]
- Found that for Linear, chunks per sync of min(20, total_chunks_worker)
works well
- Found that for Ring, chunks per sync of min(80, total_chunks_worker)
is the only scheme that beats the default and the others with any amount
of statistical significance
- Found that ring scales to 8 workers as data size increases
- Found that linear scales to 8 workers as data size increases.

Average bandwidth improvement on DiT shapes for ring is 42% and for
linear it's 70%. These are held down by regressions on very small
shapes.


### Checklist
- T3K: https://github.com/tenstorrent/tt-metal/actions/runs/17628523773
- Galaxy:
https://github.com/tenstorrent/tt-metal/actions/runs/17628512317