Skip the generated .pb.cc files from linting (including ClangTidy)
The original code was attempting to use CXX_CLANG_TIDY on a source file, but that's a target-level prop, not
a source-level prop.  So that is treated as a custom prop that we could read later, not actually doing what
was intended.  SKIP_LINTING is the source-level prop that does what we want and covers all linters, not limited
to ClangTidy.