force BH fabric to use noc1 only
force counter updates in `noc_async_write_one_packet_with_trid` for BH

We must do this because in BH, inline writes spoof their inline writes
by first writing the value to L1 and then sending a noc write out of that
L1 location to avoid a HW bug. A result of this is that those inline write
calls on BH implicitly barrier. That barrier call compares against
`noc_nonposted_writes_num_issued`, which will NOT be updated if `update_counter` is false.

Therefore, we must override `update_counter` to true if `posted` is false.

squash with noc=1 fabric on BH changes

force update counter