Enabled overriding the split reader heuristic in conv2d (#28690)
### Problem description
With the removal of `enable_split_reader` from `ttnn.Conv2dConfig` and
using the split reader heuristic
([PR](https://github.com/tenstorrent/tt-metal/pull/27731))
the usage of `ttnn.conv2d` became simpler, but the heuristic itself is
not perfect and relies on bandwidth values
calculated on TT hardware. When this code is used on custom IP, the
heuristic can be off by a wide mark
which could be detrimental performance wise, so an option to override it
is needed.

### What's changed
- added optional boolean `force_split_reader` member of
`ttnn.Conv2dConfig`, which when set overrides the split_reader
heuristic, and if not set defaults to the heuristic

### Checklist
- [X] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/17794295225)
- [X] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/17794309829)