[Fabric] Try re-enabling 1D Mux tests in CI (#28887)
### Ticket
Closes https://github.com/tenstorrent/tt-metal/issues/25288

### Problem description
The test was previously disabled because of some bug around inline
writes in BH

### What's changed
A lot of fixes have gone in, in the past couple of days and looks like
the bug is no longer reproducible. Ran the problematic test suite in a
loop locally and didnt see any issues, hence enabling it back.

### Checklist
- [x] [BH multi-card]
(https://github.com/tenstorrent/tt-metal/actions/runs/17844929623)