fix bug that disabled two txq mode in Fabric on Blackhole (#28040)
Found by inspection

The Ethernet subsystem txq and rxq assignments were being overridden to
both be 0, even though earlier code paths enabled them to be 0 and 1
(for sender channel and receiver channel out traffic, respectively).
Additionally, there was a bug with the receiver channel ack (to sender)
free running counters which is resolved. The problem was the receiver
was not keeping separate counters for each sender channel, but now it
is. Note that the free-running counter credits code path was only
enabled conditionally when two txqs are enabled, which is why this was
not observed previously.

As a part of this fix, an update to the address map for the router was
made to pack the counters contiguously in L1, indexed by channel. Each
next category of counter (e.g. ack vs completion) is aligned to the next
Ethernet word (aka 16B) aligned address.

Requires https://github.com/tenstorrent/tt-metal/pull/28801 and its
dependencies to be merged due to various stability fixes.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes:
https://github.com/tenstorrent/tt-metal/actions/runs/17898499811
- [x] BH Post Commit:
https://github.com/tenstorrent/tt-metal/actions/runs/17897607321
- [x] BH Nightly:
https://github.com/tenstorrent/tt-metal/actions/runs/17897608165
- [x] BH Multi Card Tests:
https://github.com/tenstorrent/tt-metal/actions/runs/17897611800