Fix didt test for specific chips by using submeshes (#28914)
The test was using outdated API which was broken.
The fix changes the behaviour of the specific chip test, by creating 1x1
submeshes and then assigning each logical chip as as a parameter of the
test.

This test is not part of CI pipelines, so no pipelines need to be run.