sweep dst_index for matmul API (only math core) (#604)
### Ticket
<!-- Link to Github Issue -->

https://github.com/orgs/tenstorrent/projects/123?pane=issue&itemId=118348523&issue=tenstorrent%7Ctt-llk%7C453
### Problem description
<!-- Provide context for the problem. -->
* Sweep dst_index parameter.
* Instead of performing a full sweep of the dst_index for each result's
tile_size, we test only the edge cases—specifically, the first and last
possible index values in the destination register. This approach avoids
inflating the number of test cases to 130K without gaining coverage of
new code areas or branches.
### What's changed
<!-- Describe the approach used to solve the problem.
Summarize the changes made and its impact. -->
Test cases for math_matmul --> 47232
### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update