Make size of dprint buffer core-specific (#28872)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/28871

### Problem description
Dprint buffers have a fixed size of 5 buffers per core. We don't need
that many on eth cores. The size should be made core-specific. This is
possible as dev_msgs.h is now hidden behind the HAL.

### What's changed
- Renamed PROFILER_RISC_COUNT to PROCESSOR_COUNT. This value is per core
and is used for the size of both profiler (already using it) and dprint
buffers (this PR).
- Deleted NUM_PROGRAMMABLE_CORE_TYPES and NUM_PROCESSORS_PER_CORE_TYPE.
Use the constants in core_config.h instead, which are arch-specific.
This means launch messages, etc can have different layout for different
architectures.

Mailbox size comparison (Wormhole)
| | main | this PR |
|--------|--------|--------|
| Tensix | 12768 | 12752 |
| Eth | 4576 | 3744 |

Note: it's possible to make launch and watcher msgs core-specific, but
the space saving is not that much, and currently there are some
assumptions in dispatch that size of launch messages is the same for all
cores. So, those still use the max num of processors in this arch.

Note 2: With #18742, we may be able to use only a single dprint buffer
for all cores, but Wormhole will still have separate buffers, and this
PR applies.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/17871455960)
CI passes (note: t3k fabric tests timeout is a known flakiness on main)
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/17871459661)
(timeout in blackhole-multi-card-post-commit-tests is likely unrelated)
- [x] [T3K profiler
tests](https://github.com/tenstorrent/tt-metal/actions/runs/17864589459)