remove less performant e2e variants of resnet from CI (#29123)
### Problem description


https://github.com/tenstorrent/tt-metal/actions/runs/17933988583/job/51012927955

We run 4 variants of e2e tests for Resnet50 on BH - without e2e perf
features, and with trace and/or 2cqs. The most performant version is the
one with both features on, so we should measure this time only. The
version without any of these features can be flaky, so we get fails on
CI.

For WH, we already run just the most performant and stable variant. 

### What's changed
- Bump target to account for flakiness. 
- Remove 3 less performant variants from the CI.

### Checklist

- [x] Model perf pipeline:
https://github.com/tenstorrent/tt-metal/actions/runs/17939196464