[tt-train] Add wrapper around setFabricConfig (#29120)
### Problem description
We need to pass MGD in case of LoudBox and Galaxy. In most cases it is
pretty straightforward. In perfect world, user of fabric should not be
aware about it except special cases.

### What's changed
If TT_MESH_GRAPH_DESC variable is not set -> set it based on number of
devices.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/17952891584)
CI passes
- [x] New/Existing tests provide coverage for changes