Fix llama 90b demo test in CI (#29115)
### Ticket
Link to Github Issue

### Problem description
Llama3.2-90B test appears hanging in CI. The problem is that we are
running the non-traced test case and it is indeed quite slow would
appear to be hung in the perspective of `pytest` running on some hosts.

### What's changed
Fix the traced test case to run in CI instead which is much faster than
the non-traced test case.

### Checklist
- [x] [T3000 demo
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-demo-tests.yaml)
CI passes
  - https://github.com/tenstorrent/tt-metal/actions/runs/17948178767
- [x] revert 4ec942e3c3ca4cccfccd9adf61fe8ff482c48846
- through
[84bc10a](https://github.com/tenstorrent/tt-metal/pull/29115/commits/84bc10a8aae3e7beab4a6293645a51f47fca3036)