conv3.0 batch fix (#28798)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/28386

### Problem description
When a single core has to process more than 1 batch, once it is done
with the previous batch, we need to 'restart' the optimization and fill
in the whole CB before starting to reuse data again.

### What's changed
- Added a compile time arg to know whether we need to check for
situations like this one (to avoid additional if-s in the code in case
we know that no core will process more than 1 batch.
- When a new reader indices segment is encountered, we check if the new
start index is further than expected from the previous end index - if
so, we need to initialize the reuse again and fill in the whole CB
- Refactored out repeated code to `load_new_segment` function

Note 1: This means that `we determine on-flight whether we need to
restart the reuse` - each time we load a new segment. Another option was
to calculate this on host and pass to each core runtime info on where to
expect the new batch starts, however, this has a potential to explode as
there is no limit on how many batches to expect on one core. Moreover,
it would require extra if-s in the code as frequent as we do with the
proposed solution in the PR.

Note 2: The 'restart' means that we:
1. finish processing the pixels from the current image width without any
reuse
2. use the next whole image width to fill in the CB and 'initialize' the
optimization
3. start the reuse from the next image width

This means that we accept the fact that we didn't maximize the reuse -
point 1; we essentially could start the new image width right when we
encounter new batch, however, this would mean that compute kernel would
need to be aware of this and update the reader pointers accordingly, and
I chose to avoid this.

Note 3: We already had tests with batch > 1 in the suite, but the PCC
check wasn't strict enough - this means the output could be bad only on
a couple of rows of couple of cores. Updated the PCC to be 0.999 - the
test would fail this check before the fix.

### Checklist

- [x] Nightly l2:
https://github.com/tenstorrent/tt-metal/actions/runs/17860097848
- [x] APC:
https://github.com/tenstorrent/tt-metal/actions/runs/17860088099
- [x] BH post commit (p100):
https://github.com/tenstorrent/tt-metal/actions/runs/17860093781
- [x] Device perf:
https://github.com/tenstorrent/tt-metal/actions/runs/17860083969