[skip ci] ci: format test results in PR comment (#695)
### Ticket
None

### Problem description
Presented test results could be seen as confusing, as L2 APC nightly
name is not clear.

### What's changed
This PR improves the clarity of test results displayed in PR comments by
renaming the confusing "L2 Nightly APC" label to a more descriptive
"Wormhole C++ tests" and reordering the display of test results.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update