Increased timeout for L2 nightly tests (#29431)
### Problem description
Nightly L2 pipeline being red on main due to timeout triggered. Conv
tests were actually running that long, with the reasonable amount per
test so the investigation showed that it wasn't due to the hang.

### What's changed
To unblock the pipeline I increased timeout for nightly pipeline for
now. Conv test clean up is planned.