Disable noc flush for inline writes to L1 in fabric router (#29331)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-metal/issues/28758)

### Problem description
To avoid a blackhole bug when using inline writes, we issue normal async
write instead of inline writes to a dest L1, but we need to have a noc
flush to avoid src address being overwritten. This is causing perf
degradation is BH

### What's changed
on the fabric router side, add a new dedicated address for src address
update, and when update the read counter for connected workers, just
write to that location without the need for flush.

This is fine since counters are free counter, even if a new counter
value overwrite that location before the previous txn is flushed, it
just means worker sees a newer value twice, and this is not changing
anything.

### Checklist
- [x] [All post commit]
https://github.com/tenstorrent/tt-metal/actions/runs/18018486984/job/51270553639
- [x] [Blackhole Post commit]
https://github.com/tenstorrent/tt-metal/actions/runs/18020152169
- [x] BH nightly
https://github.com/tenstorrent/tt-metal/actions/runs/18020155698/job/51276243847
- [x] t3000 unit
https://github.com/tenstorrent/tt-metal/actions/runs/18020162777/job/51276261389