#14356: Add nan handling for acos and asin (#29215)
### Ticket
Link to Github Issue #14356 

### Problem description
If Input value is not within range [-1,1] , it should produce nan as
result. This case is not handled in the ckernel file for both ttnn.acos
and ttnn.asin.

### What's changed
Updated the llk to handle this case.

### Perf Results :

**Updated version:**
asin -2772 ns (~10% slower)
acos - 2876 ns (~2% slower)

**Older version:**
asin - 2512 ns
acos -  2805 ns

### Note :
The valid range for acos and asin is -1 to +1. Within -0.9 to +0.9, the
ULP stays below 5, but outside this range it increases significantly —
around 30 for asin and 65 for acos

### Checklist
- [ ] All post commit