initialize CB wr pointer at the beginning of each act block h (#29239)
### Ticket
#29005

### Problem description
Watcher asserted in conv with activation reuse when ran on BH.

### What's changed
Beacuse of the optimization, we need to set the write address ourself
and not rely on cb wr pointer.
When doing a new act block h, we reinitialize the write address to the
beginning of the CB, but we don't update hte CB wr pointer as well so
push back at some points sets write pointer to an address beyound limit.
Since we don't use the write ptr for actual data write, this didn't
result in PCC issues.

### Checklist

- [x] APC:
https://github.com/tenstorrent/tt-metal/actions/runs/18099934758
- [x] BH post commit:
https://github.com/tenstorrent/tt-metal/actions/runs/18099928891 (LLM
box failing as on main)
- [x] Device perf:
https://github.com/tenstorrent/tt-metal/actions/runs/18099939334 (swin
v2 failing on main, PR to fix in progress)
- [x] Nightly L2:
https://github.com/tenstorrent/tt-metal/actions/runs/18099915046