Change tensor page size in argmax single core kernel (#29414)
### Ticket
[23001](https://github.com/tenstorrent/tt-metal/issues/23001)
[23000](https://github.com/tenstorrent/tt-metal/issues/23000)

### Problem description
The argmax operation generates wrong output in several argmax sweep test
cases.
The root cause of the problem is that invalid tensor page sizes are set
in the program factory.
The kernel code iterates over elements of the output tensor: for each
output element, all required input elements are read as a single page,
then index of the maximal element is found and written to the output
tensor also as a single page. What follows, the kernel code requires the
src page size to be the size of elements in the reduction dimension, and
the dst page size equal to the size of a single output element. The
program factory code applied round_up_to_mul32() to the (otherwise
correct) page sizes, which was then not handled in the kernel code.

### What's changed
The page sizes are changed (the rounding is removed) to reflect the
tensor access pattern in the kernel code.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI with demo tests passes (if applicable)
- [x] [Model
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml)
CI passes (if applicable)