Removing capnp EzRpcServer because of the bug (#29720)
During introduction of capnp, we hit a 12 year old bug that we
discovered and reported here:
https://github.com/capnproto/capnproto/issues/2418#issuecomment-3349246495.
Here we replace `EzRpcServer` with `TwoPartyServer` and removing capnp
patch that was mitigating original bug.
Usage of `TwoPartyServer` allowed us to expose feature of using unix
sockets (adding test to verify that it works).